Restructure test file structure
All utility .c and .h files have been moved to new utils/ dir, src/ has been renamed to unit_tests/, and test_real_data.c has been moved to root tests/
