+++ dev/CMakeLists.txt
@@ -1,6 +1,6 @@
 if(ASAN)
-  message(STATUS "Building test_asan...")
-  add_executable(test_asan test.c ../tests/src/io.c)
+  message(STATUS "Building executable test_asan")
+  add_executable(test_asan test.c ../tests/utils/io.c)
   set_target_properties(test_asan PROPERTIES
     EXCLUDE_FROM_ALL $<NOT:$<CONFIG:Debug>>
   )
@@ -8,7 +8,7 @@ if(ASAN)
   target_link_libraries(test_asan openblas)
   target_link_libraries(test_asan hmat_lib)
 else()
-  message(STATUS "Building test_unsafe...")
+  message(STATUS "Building executable test_unsafe")
   add_executable(test_unsafe test.c ../tests/src/io.c)
   set_target_properties(test_unsafe PROPERTIES
     EXCLUDE_FROM_ALL $<NOT:$<CONFIG:Debug>>
