+++ src/CMakeLists.txt
@@ -25,6 +25,7 @@ if(OpenMP_C_FOUND)
 endif()
 
 if(ASAN)
+  message(STATUS "Building library with ASAN ON")
   if (CMAKE_C_COMPILER_ID MATCHES "Clang|GNU")
     add_compile_options($<$<CONFIG:Debug>:-fsanitize=address -fsanitize=leak -fsanitize=undefined>)
     add_link_options($<$<CONFIG:Debug>:-fsanitize=address -fsanitize=leak -fsanitize=undefined>)
@@ -39,5 +40,7 @@ if(ASAN)
   else()
     message(WARNING "Address Sanitizer not available for ${CMAKE_C_COMPILER_ID}")
   endif()
+else()
+  message(STATUS "Building library with ASAN OFF")
 endif()
