+++ tests/CMakeLists.txt
@@ -1,48 +1,18 @@
+add_subdirectory(unit_tests)
+
 if(BUILD_TESTING)
-  add_executable(
-    unit_test 
-    src/test_allocators.c src/test_tree_construct.c src/test_vector_algebra.c 
-    src/test_hodlr_dense_algebra.c src/test_hodlr_hodlr_algebra.c
-    ../src/lapack_wrapper.c src/utils.c 
-    src/common_data.c ../src/utils.c
-  )
-    
-  target_link_libraries(unit_test ${MATH_LIBRARY})
-  target_link_libraries(unit_test openblas)
-  target_link_libraries(unit_test criterion)
-
-  add_test(NAME unit_test COMMAND unit_test -j1 --verbose=0)
-
-  set_property(TEST unit_test PROPERTY
-      ENVIRONMENT "CTEST_OUTPUT_ON_FAILURE=1"
-  )
-
-  if(BUILD_OPENMP_TESTS AND OpenMP_C_FOUND)
-    add_executable(
-      test_constructors_omp 
-      src/test_allocators.c src/test_tree_construct.c ../src/lapack_wrapper.c 
-      src/utils.c src/common_data.c ../src/utils.c
-    )
-
-    target_link_libraries(test_constructors_omp ${MATH_LIBRARY})
-    target_link_libraries(test_constructors_omp openblas)
-    target_link_libraries(test_constructors_omp criterion)
-    target_link_libraries(test_constructors_omp OpenMP::OpenMP_C)
-
-    add_test(NAME test_constructors_omp COMMAND test_constructors_omp -j1 --verbose=0)
-
-    set_property(TEST test_constructors_omp PROPERTY
-        ENVIRONMENT "CTEST_OUTPUT_ON_FAILURE=1"
-    )
-  endif()
-
   if(BUILD_REAL_DATA_TESTS)
+    message(STATUS "Building tests that use real-world data")
+
     add_executable(
       test_real_data 
-      src/test_real_data.c src/utils.c src/common_data.c src/io.c 
-      ../dev/common.c ../src/allocators.c ../src/constructors.c 
-      ../src/lapack_wrapper.c ../src/vector_algebra.c ../src/dense_algebra.c
-      ../src/utils.c ../src/hodlr_algebra.c
+      ./test_real_data.c 
+      utils/utils.c utils/common_data.c utils/io.c 
+      ../dev/common.c 
+      ../src/allocators.c ../src/constructors.c 
+      ../src/vector_algebra.c ../src/dense_algebra.c
+      ../src/hodlr_algebra.c
+      .../src/utils.c ./src/lapack_wrapper.c 
     )
 
     target_link_libraries(test_real_data ${MATH_LIBRARY})
@@ -50,10 +20,12 @@ if(BUILD_TESTING)
     target_link_libraries(test_real_data criterion)
 
     if(BUILD_OPENMP_TESTS AND OpenMP_C_FOUND)
+      message(STATUS "Building real-world data tests with OpenMP")
       target_link_libraries(test_real_data OpenMP::OpenMP_C)
     endif()
 
     if(REAL_DATA_PRINT_S)
+      message(STATUS "Real-world data tests will log ranks")
       target_compile_definitions(test_real_data PRIVATE HODLR_REAL_DATA_PRINT_S)
     endif()
     
@@ -71,6 +43,4 @@ if(BUILD_TESTING)
       SYMBOLIC
     )
   endif()
-
-  add_custom_target(vtest COMMAND ${CMAKE_CTEST_COMMAND} --verbose)
 endif()
