+++ tests/unit_tests/test_hodlr_hodlr_algebra.c
@@ -11,8 +11,8 @@
 #include <criterion/new/assert.h>
 #include <criterion/logging.h>
 
-#include "../include/utils.h"
-#include "../include/common_data.h"
+#include "../utils/utils.h"
+#include "../utils/common_data.h"
 
 #include "../../include/hmat_lib/utils.h"
