+++ tests/unit_tests/test_tree_construct.c
@@ -11,8 +11,8 @@
 #include <criterion/new/assert.h>
 #include <criterion/logging.h>
 
-#include "../include/utils.h"
-#include "../include/common_data.h"
+#include "../utils/utils.h"
+#include "../utils/common_data.h"
 
 #include "../../src/constructors.c"
 #include "../../include/hmat_lib/error.h"
