+++ tests/utils/utils.h
@@ -138,3 +138,4 @@ void copy_block_sizes(
   struct TreeHODLR *dest,
   const bool copy_s
 );
+
